PTDERRRRR JAI RM -RF

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: jprofit <jprofit@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 17:16:02 by jprofit           #+#    #+#             */
-/*   Updated: 2022/11/14 09:19:21 by jprofit          ###   ########.fr       */
+/*   Updated: 2022/11/17 15:30:59 by jprofit          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
